Refactor Docker and Kubernetes deploy backends
