+++ nvflare/tool/deploy/templates/docker/start_docker.sh
@@ -1,0 +1,40 @@
+#!/usr/bin/env bash
+# Generated by NVFlare deploy prepare - Docker runtime @@NVFLARE_ROLE_LABEL@@ startup script.
+# The parent @@NVFLARE_ROLE_LABEL@@ container mounts this prepared kit and launches job containers through Docker.
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+HOST_WORKSPACE="$(cd "$DIR/.." && pwd)"
+DOCKER_IMAGE=${NVFL_P_IMAGE:-@@NVFLARE_DOCKER_IMAGE@@}
+
+if ! docker info > /dev/null 2>&1; then
+    echo "ERROR: cannot connect to Docker daemon. Make sure your user has permission to run docker."
+    echo "  Option 1 (docker group): sudo usermod -aG docker \$USER  then log out and back in."
+    echo "  Option 2 (rootless Docker): https://docs.docker.com/engine/security/rootless/"
+    exit 1
+fi
+
+echo "Starting NVFlare @@NVFLARE_ROLE_LABEL@@ @@NVFLARE_SITE_NAME@@ with image $DOCKER_IMAGE"
+echo "Host workspace: $HOST_WORKSPACE"
+
+NETWORK_NAME=@@NVFLARE_NETWORK_NAME@@
+if ! docker network ls --filter name="$NETWORK_NAME" --format "{{.Name}}" | grep -wq "$NETWORK_NAME"; then
+    docker network create "$NETWORK_NAME"
+    echo "Created Docker network: $NETWORK_NAME"
+fi
+
+rm -f "$HOST_WORKSPACE/daemon_pid.fl"
+
+SOCK_GID=$(stat -c '%g' /var/run/docker.sock 2>/dev/null || stat -f '%g' /var/run/docker.sock 2>/dev/null || echo "")
+GROUP_ADD_ARG=""
+if [ -n "$SOCK_GID" ] && [ "$SOCK_GID" != "0" ]; then
+    GROUP_ADD_ARG="--group-add $SOCK_GID"
+fi
+
+docker run --name @@NVFLARE_CONTAINER_NAME@@ \
+    --user "$(id -u):$(id -g)" \
+    $GROUP_ADD_ARG \
+    --network "$NETWORK_NAME" \
+@@NVFLARE_NETWORK_ALIAS@@    -v "$HOST_WORKSPACE":@@NVFLARE_WORKSPACE_MOUNT_PATH@@ \
+    -v /var/run/docker.sock:/var/run/docker.sock \
+    -e NVFL_DOCKER_WORKSPACE="$HOST_WORKSPACE" \
+@@NVFLARE_PUBLISH_PORTS@@    --rm "$DOCKER_IMAGE" \
+    @@NVFLARE_PARENT_COMMAND@@
