+++ tests/unit_test/tool/deploy/deploy_commands_test.py
@@ -14,6 +14,7 @@
 
 import argparse
 import base64
+import hashlib
 import json
 import shutil
 import subprocess
@@ -26,17 +27,9 @@ from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.x509.oid import NameOID
 
-from nvflare.tool.deploy.deploy_commands import (
-    GPU_RESOURCE_CONSUMER,
-    GPU_RESOURCE_MANAGER,
-    HELM_RELEASE_NAME_MAX_LENGTH,
-    K8S_PARENT_PYTHON_PATH,
-    PROCESS_CLIENT_LAUNCHER,
-    _k8s_release_name,
-    prepare_deployment,
-    stage_k8_deployment,
-    unstage_k8_deployment,
-)
+from nvflare.tool.deploy.deploy_commands import prepare_deployment, stage_k8_deployment, unstage_k8_deployment
+from nvflare.tool.deploy.deploy_common import GPU_RESOURCE_CONSUMER, GPU_RESOURCE_MANAGER, PROCESS_CLIENT_LAUNCHER
+from nvflare.tool.deploy.k8s_deploy import HELM_RELEASE_NAME_MAX_LENGTH, K8S_PARENT_PYTHON_PATH, _k8s_release_name
 
 
 def _write_json(path, data):
@@ -210,7 +203,7 @@ def _capture_kubectl(monkeypatch, returncode=0, fail_on=None):
             return subprocess.CompletedProcess(cmd, 1, stdout="", stderr="boom")
         return subprocess.CompletedProcess(cmd, returncode, stdout="ok", stderr="")
 
-    monkeypatch.setattr("nvflare.tool.deploy.deploy_commands.subprocess.run", fake_run)
+    monkeypatch.setattr("nvflare.tool.deploy.k8s_stage.subprocess.run", fake_run)
     return calls
 
 
@@ -268,7 +261,13 @@ def test_prepare_docker_client_copies_and_patches_runtime_files(tmp_path, capsys
     assert not (output / "startup" / "sub_start.sh").exists()
     assert not (output / "startup" / "stop_fl.sh").exists()
     assert not (output / "startup" / "docker.sh").exists()
-    script = (output / "startup" / "start_docker.sh").read_text()
+    script_path = output / "startup" / "start_docker.sh"
+    script_bytes = script_path.read_bytes()
+    assert (
+        hashlib.sha256(script_bytes).hexdigest() == "592e3a3bba65d9e0f446e793a60621545d1c6dc022241205f6ff0831692c1679"
+    )
+    assert script_path.stat().st_mode & 0o777 == 0o755
+    script = script_bytes.decode()
     assert "repo/nvflare:dev" in script
     assert 'NETWORK_NAME="nvflare-test"' in script
     assert "--network-alias" not in script
@@ -296,7 +295,11 @@ def test_prepare_docker_client_copies_and_patches_runtime_files(tmp_path, capsys
 
     comm_config = json.loads((output / "local" / "comm_config.json").read_text())
     assert comm_config["internal"]["resources"]["host"] == "0.0.0.0"
-    study_runtime = yaml.safe_load((output / "local" / "study_runtime.yaml").read_text())
+    study_runtime_path = output / "local" / "study_runtime.yaml"
+    assert hashlib.sha256(study_runtime_path.read_bytes()).hexdigest() == (
+        "b12fee67d4992e9aeeda532f8cdc9988983bb43c716f5cb9dde4b2076b2c2950"
+    )
+    study_runtime = yaml.safe_load(study_runtime_path.read_text())
     assert study_runtime == {"format_version": 2, "studies": {}}
     assert not (output / "local" / "study_data.yaml").exists()
 
@@ -1164,7 +1167,7 @@ def test_stage_k8_reports_kubectl_launch_os_error(tmp_path, capsys, monkeypatch)
     def fake_run(_cmd, **_kwargs):
         raise PermissionError("denied")
 
-    monkeypatch.setattr("nvflare.tool.deploy.deploy_commands.subprocess.run", fake_run)
+    monkeypatch.setattr("nvflare.tool.deploy.k8s_stage.subprocess.run", fake_run)
 
     with pytest.raises(SystemExit):
         stage_k8_deployment(_stage_k8_args(output, namespace="nvflare"))
